[FIX] Fixed behavior when adding/removing items
There was a typo so that prepending items was not working
properly. It has been fixed to call the right method.

Removing an item should only work when the paPackery controller is
not being destroyed. The logic was the inverse of what it was
supposed to be and could lead to memory leaks. The condition check
when removing items has been reversed to have the right behavior.